********************************************************** ** Big Maps Projects code (incl. Multiplayer v1.5 ** ********************************************************** - Merged Big Maps Project code from BMP+MP trunk (Revision: 3340) o Complete SVN Revision history: https://81.169.133.124/source/ja2/branches/Wanne/JA2%201.13%20MP - Before THIS merge, I made a branch of the existing 1.13 source o SVN Branch: https://81.169.133.124/source/ja2/branches/JA2_rev.3336/src - Removed old VS 6.0 and VS 2003 project and solutions files, because compilation is broken long time ago - I will add VS 2010 projects and solution file in the next few days
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@3341 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Editor/Editor Modes.cpp
@@ -256,7 +256,7 @@ void SetEditorTerrainTaskbarMode( UINT16 usNewMode )
 
 void ShowExitGrids()
 {
-	UINT16 i;
+	INT32 i;
 	LEVELNODE *pLevelNode;
 	if( gfShowExitGrids )
 		return;
@@ -272,7 +272,7 @@ void ShowExitGrids()
 
 void HideExitGrids()
 {
-	UINT16 i;
+	INT32 i;
 	LEVELNODE *pLevelNode;
 	if( !gfShowExitGrids )
 		return;
