+++ Editor/Smoothing Utils.h
@@ -38,28 +38,28 @@ enum {
 
 //in newsmooth.c
 extern INT8 gbWallTileLUT[NUM_WALL_TYPES][7];
-extern void EraseWalls( UINT32 iMapIndex );
-extern void BuildWallPiece( UINT32 iMapIndex, UINT8 ubWallPiece, UINT16 usWallType );
+extern void EraseWalls( INT32 iMapIndex );
+extern void BuildWallPiece( INT32 iMapIndex, UINT8 ubWallPiece, UINT16 usWallType );
 //in Smoothing Utils
-void RestoreWalls( UINT32 iMapIndex );
-UINT16 SearchForRoofType( UINT32 iMapIndex );
-UINT16 SearchForWallType( UINT32 iMapIndex );
-BOOLEAN RoofAtGridNo( UINT32 iMapIndex );
-BOOLEAN BuildingAtGridNo( UINT32 iMapIndex );
-LEVELNODE* GetHorizontalWall( UINT32 iMapIndex );
-LEVELNODE* GetVerticalWall( UINT32 iMapIndex );
-LEVELNODE* GetVerticalFence( UINT32 iMapIndex );
-LEVELNODE* GetHorizontalFence( UINT32 iMapIndex );
-UINT16 GetHorizontalWallType( UINT32 iMapIndex );
-UINT16 GetVerticalWallType( UINT32 iMapIndex );
-void EraseHorizontalWall( UINT32 iMapIndex );
-void EraseVerticalWall( UINT32 iMapIndex );
-void ChangeHorizontalWall( UINT32 iMapIndex, UINT16 usNewPiece );
-void ChangeVerticalWall( UINT32 iMapIndex, UINT16 usNewPiece );
+void RestoreWalls( INT32 iMapIndex );
+UINT16 SearchForRoofType( INT32 iMapIndex );
+UINT16 SearchForWallType( INT32 iMapIndex );
+BOOLEAN RoofAtGridNo( INT32 iMapIndex );
+BOOLEAN BuildingAtGridNo( INT32 iMapIndex );
+LEVELNODE* GetHorizontalWall( INT32 iMapIndex );
+LEVELNODE* GetVerticalWall( INT32 iMapIndex );
+LEVELNODE* GetVerticalFence( INT32 iMapIndex );
+LEVELNODE* GetHorizontalFence( INT32 iMapIndex );
+UINT16 GetHorizontalWallType( INT32 iMapIndex );
+UINT16 GetVerticalWallType( INT32 iMapIndex );
+void EraseHorizontalWall( INT32 iMapIndex );
+void EraseVerticalWall( INT32 iMapIndex );
+void ChangeHorizontalWall( INT32 iMapIndex, UINT16 usNewPiece );
+void ChangeVerticalWall( INT32 iMapIndex, UINT16 usNewPiece );
 UINT16 GetWallClass( LEVELNODE *pWall );
-UINT16 GetVerticalWallClass( UINT16 iMapIndex );
-UINT16 GetHorizontalWallClass( UINT16 iMapIndex );
-BOOLEAN ValidDecalPlacement( UINT32 iMapIndex );
+UINT16 GetVerticalWallClass( INT32 iMapIndex );
+UINT16 GetHorizontalWallClass( INT32 iMapIndex );
+BOOLEAN ValidDecalPlacement( INT32 iMapIndex );
 
 #endif
 #endif
