+++ Strategic/Queen Command.h
@@ -54,7 +54,7 @@ BOOLEAN PlayerSectorDefended( UINT8 ubSectorID );
 
 BOOLEAN OnlyHostileCivsInSector();
 
-extern INT16 gsInterrogationGridNo[3];
+extern INT32 gsInterrogationGridNo[3];
 
 BOOLEAN CheckPendingEnemies();
