+++ Strategic/Quest Debug System.cpp
@@ -477,7 +477,7 @@ SCROLL_BOX	gItemListBox;				// The Npc Scroll box
 
 SCROLL_BOX	*gpActiveListBox;		// Only 1 scroll box is active at a time, this is set to it.
 
-INT16				gsQdsEnteringGridNo	=0;
+INT32				gsQdsEnteringGridNo	=0;
 
 
 UINT8				gubTextEntryAction = QD_DROP_DOWN_NO_ACTION;
@@ -3008,7 +3008,7 @@ void AddNPCToGridNo( INT32 iGridNo )
 	MercCreateStruct.sSectorX			= sSectorX;
 	MercCreateStruct.sSectorY			= sSectorY;
 	MercCreateStruct.bSectorZ			= gbWorldSectorZ;
-	MercCreateStruct.sInsertionGridNo		= (INT16) iGridNo;
+	MercCreateStruct.sInsertionGridNo		= iGridNo;
 
 //	RandomizeNewSoldierStats( &MercCreateStruct );
 
@@ -3023,14 +3023,14 @@ void AddNPCToGridNo( INT32 iGridNo )
 	//Add all the npc in the current sectory the npc array
 	AddNPCsInSectorToArray();
 
-	gsQdsEnteringGridNo = (INT16)iGridNo;
+	gsQdsEnteringGridNo = iGridNo;
 }
 
 
 
 void AddItemToGridNo( INT32 iGridNo )
 {
-	gsQdsEnteringGridNo = (INT16)iGridNo;
+	gsQdsEnteringGridNo = iGridNo;
 
 
 	if( Item[ gItemListBox.sCurSelectedItem ].usItemClass == IC_KEY )
