+++ Strategic/Quest Debug System.h
@@ -6,7 +6,7 @@
 
 
 extern	BOOLEAN			gfNpcLogButton;
-extern	INT16				gsQdsEnteringGridNo;
+extern	INT32				gsQdsEnteringGridNo;
 
 
 void NpcRecordLoggingInit( UINT8 ubNpcID, UINT8 ubMercID, UINT8 ubQuoteNum, UINT8 ubApproach );
