+++ Strategic/Quests.cpp
@@ -185,7 +185,7 @@ BOOLEAN CheckGuyVisible( UINT8 ubNPC, UINT8 ubGuy )
 	}
 }
 
-BOOLEAN CheckNPCAt( UINT8 ubNPC, INT16 sGridNo )
+BOOLEAN CheckNPCAt( UINT8 ubNPC, INT32 sGridNo )
 {
 	SOLDIERTYPE * pNPC;
 
@@ -227,7 +227,7 @@ BOOLEAN CheckNPCIsEnemy( UINT8 ubProfileID )
 BOOLEAN CheckIfMercIsNearNPC( SOLDIERTYPE *pMerc, UINT8 ubProfileId )
 {
 	SOLDIERTYPE *		pNPC;
-	INT16						sGridNo;
+	INT32 sGridNo;
 
 	// no merc nearby?
 	if ( pMerc == NULL )
@@ -258,7 +258,7 @@ INT8 NumWoundedMercsNearby( UINT8 ubProfileID )
 	UINT32					uiLoop;
 	SOLDIERTYPE *		pNPC;
 	SOLDIERTYPE *		pSoldier;
-	INT16						sGridNo;
+	INT32 sGridNo;
 
 	pNPC = FindSoldierByProfileID( ubProfileID, FALSE );
 	if (!pNPC)
@@ -289,7 +289,7 @@ INT8 NumMercsNear( UINT8 ubProfileID, UINT8 ubMaxDist )
 	UINT32					uiLoop;
 	SOLDIERTYPE *		pNPC;
 	SOLDIERTYPE *		pSoldier;
-	INT16						sGridNo;
+	INT32 sGridNo;
 
 	pNPC = FindSoldierByProfileID( ubProfileID, FALSE );
 	if (!pNPC)
@@ -436,7 +436,7 @@ INT8 NumMalesPresent( UINT8 ubProfileID )
 	UINT32					uiLoop;
 	SOLDIERTYPE *		pNPC;
 	SOLDIERTYPE *		pSoldier;
-	INT16						sGridNo;
+	INT32 sGridNo;
 
 	pNPC = FindSoldierByProfileID( ubProfileID, FALSE );
 	if (!pNPC)
@@ -470,7 +470,7 @@ BOOLEAN FemalePresent( UINT8 ubProfileID )
 	UINT32					uiLoop;
 	SOLDIERTYPE *		pNPC;
 	SOLDIERTYPE *		pSoldier;
-	INT16						sGridNo;
+	INT32 sGridNo;
 
 	pNPC = FindSoldierByProfileID( ubProfileID, FALSE );
 	if (!pNPC)
@@ -547,7 +547,7 @@ BOOLEAN CheckNPCSector( UINT8 ubProfileID, INT16 sSectorX, INT16 sSectorY, INT8 
 
 }
 
-BOOLEAN AIMMercWithin( INT16 sGridNo, INT16 sDistance )
+BOOLEAN AIMMercWithin( INT32 sGridNo, INT16 sDistance )
 {
 	UINT32					uiLoop;
 	SOLDIERTYPE *		pSoldier;
@@ -1250,11 +1250,11 @@ void InternalStartQuest( UINT8 ubQuest, INT16 sSectorX, INT16 sSectorY, BOOLEAN 
 	{
 		gubQuest[ubQuest] = QUESTINPROGRESS;
 
-		if ( fUpdateHistory )
-		{
+	if ( fUpdateHistory )
+	{
 			if (!is_networked)
-				SetHistoryFact( HISTORY_QUEST_STARTED, ubQuest, GetWorldTotalMin(), sSectorX, sSectorY );
-		}
+		SetHistoryFact( HISTORY_QUEST_STARTED, ubQuest, GetWorldTotalMin(), sSectorX, sSectorY );
+	}
 	}
 	else
 	{
@@ -1333,12 +1333,11 @@ void CheckForQuests( UINT32 uiDay )
 	// already started
 	if (gubQuest[QUEST_DELIVER_LETTER] == QUESTNOTSTARTED)
 	{
-
 		StartQuest( QUEST_DELIVER_LETTER, -1, -1 );
 #ifdef TESTQUESTS
 
 		if (!is_networked)
-			ScreenMsg( MSG_FONT_RED, MSG_DEBUG, L"Started DELIVER LETTER quest");
+		ScreenMsg( MSG_FONT_RED, MSG_DEBUG, L"Started DELIVER LETTER quest");
 #endif
 	}
